努力ノートの改善web
