+++ reactjs/src/native/components/molecules/challenges/ChallengeNoteForm.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ChallengeNoteForm = (props: any) => {
+  return null;
+};
+
+export default ChallengeNoteForm;
