+++ reactjs/src/native/containers/challenges/ChallengeNoteFormContainer.tsx
@@ -1,0 +1,4 @@
+import connect from '~/connects/challenges/ChallengeNoteFormConnect';
+import ChallengeNoteForm from '~/native/components/molecules/challenges/ChallengeNoteForm';
+
+export default connect(ChallengeNoteForm);
